feat: limit setup vault name length
