+++ clients/desktop/src/i18n/locales/de.ts
@@ -328,6 +328,8 @@ export const de = {
   or: 'oder',
   enter_vault_name: 'Tresorname eingeben',
   vault_name_required: 'Tresorname ist erforderlich',
+  vault_name_max_length_error:
+    'Der Tresorname darf höchstens 50 Zeichen lang sein',
   vault_name_already_exists: 'Tresorname muss eindeutig sein',
   vault_settings_address_book_save_addresses_button: 'Adresse speichern',
   vault_settings_address_book_add_addresses_title: 'Adresse hinzufügen',
