+++ clients/desktop/src/i18n/locales/es.ts
@@ -328,6 +328,8 @@ export const es = {
   or: 'o',
   enter_vault_name: 'Ingrese el nombre de la bóveda',
   vault_name_required: 'Se requiere el nombre de la bóveda',
+  vault_name_max_length_error:
+    'El nombre del bóveda debe tener menos de 50 caracteres',
   vault_name_already_exists: 'El nombre de la bóveda debe ser único',
   vault_settings_address_book_save_addresses_button: 'Guardar dirección',
   vault_settings_address_book_add_addresses_title: 'Agregar dirección',
