+++ clients/desktop/src/vault/setup/shared/constants.ts
@@ -1,0 +1,1 @@
+export const MAX_VAULT_NAME_LENGTH = 50;
